pyproject.toml
[project]
name = "spaceguard"
version = "0.1.0"
description = "macOS menu bar utility: disk/swap monitoring and safe cache cleanup"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyside6>=6.7.0",
]

[project.scripts]
spaceguard = "spaceguard.app:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "nuitka>=2.0.0",
    "pillow>=10.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
